https://github.com/next-exp/IC/pull/807

[author: gondiaz]

This PR is intended to update the IC conda and libraries to allow working in Python 3.9 3.8.

The branch has been tested on the following systems:

- MacOS Monterey (12.0.1)
- Ubuntu 20.04 (new system for github actions)
- Ubuntu 16.04 (gpu1next)
- Red Hat Enterprise Linux Server 7.5 (Maipo) (CESGA)
- CentOS Linux 7 (majorana1)
- CentOS Linux 7 (frontend1demo)
be installed appart, ppc64el arch)
- Scientific Linux 6.6 (neutrinos1) --> failing conda installation.

Neutrinos: The version of the libc library needed (>=2.14) is not
available in neutrinos. We would need to install them separately and
we might do so in the future, but for now, we stick to 3.8.

The Ubuntu version of github actions has been updated to Ubuntu
20.04 from Ubuntu 18.04 to allow tests to pass.

Some tests might fail in some machines due to the hypothesis default
timing. That doesn't mean that the test is not passing (namely not
fullfiling the assertions) but just that hypothesis califies them as failed.

[reviewer: gonzaponte]

Upgrade-related issues addressed clearly and warnings
silenced. Hopefully, the next upgrade won't take this long :) Good job.

+++ invisible_cities/core/fit_functions.py
@@ -25,16 +25,16 @@ def fixed_parameters(fn, **kwargs):
     fn_sig    = insp.signature(fn)
     fn_pars   = np.array(list(fn_sig.parameters))[1:]
     par_gen   = (kwargs.get(k, np.nan) for k in fn_pars)
-    all_args  = np.fromiter(par_gen, np.float)
+    all_args  = np.fromiter(par_gen, float)
     free_pars = np.isnan(all_args)
-    
+
     if np.all(free_pars):
         raise ValueError(str(kwargs.keys()) + " not parameters " + fn.__name__)
     elif not np.any(free_pars):
         raise ValueError("Fixing all parameters pointless")
     elif np.count_nonzero(free_pars) > len(fn_pars) - len(kwargs):
         raise ValueError("Some parameters not found in " + fn.__name__)
-    
+
     @wraps(fn)
     def fixed_fn(x, *pars):
         all_args[free_pars] = pars
@@ -43,11 +43,11 @@ def fixed_parameters(fn, **kwargs):
         if hasattr(fn, 'n_gaussians'):
             fixed_fn.n_gaussians = fn.n_gaussians
         return func_value
-    
+
     ## Correct the doc string and signature for the new wrapped function
     fixed_fn.__doc__  = fixed_fn.__name__ + str(fixed_fn.__doc__) + "\n"
     fixed_fn.__doc__ += " fixed" + str(fn_pars[np.invert(free_pars)])
-    
+
     new_pars               = np.array(tuple(fn_sig.parameters.values()))
     new_pars               = np.insert(new_pars[1:][free_pars], 0, new_pars[0])
     fixed_fn.__signature__ = fn_sig.replace(parameters=tuple(new_pars))
